10/05 - 16:55

+++ back-end-maven/src/main/java/net/weg/wegssm/controller/AnexoController.java
@@ -5,12 +5,10 @@ import net.weg.wegssm.model.entities.Anexo;
 import net.weg.wegssm.model.service.AnexoService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import javax.transaction.Transactional;
-import java.io.IOException;
 
 /**
  * Classe Controller para as funções dos anexos
@@ -27,36 +25,34 @@ public class AnexoController {
 
     /**
      * Função para criar um anexo no banco de dados, recebendo um AnexoDTO no body
+     *
+     * @param arquivo Arquivo para salvamento
+     * @return ResponseEntity com o anexo recém-salvo
      */
     @PostMapping
-    public ResponseEntity<Anexo> save(@RequestParam(value = "anexo") MultipartFile file) throws IOException {
-        Anexo anexo = new Anexo(file.getOriginalFilename(), file.getContentType(), file.getBytes());
-        return ResponseEntity.status(HttpStatus.OK).body(anexoService.save(anexo));
+    public ResponseEntity<Object> save(@RequestParam(value = "anexo") MultipartFile arquivo) {
+        try {
+            Anexo anexo = new Anexo(null, arquivo.getOriginalFilename(), arquivo.getContentType(), arquivo.getBytes());
+            return ResponseEntity.status(HttpStatus.OK).body(anexoService.save(anexo));
+        } catch (Exception exception) {
+            return ResponseEntity.status(HttpStatus.CONFLICT).body("Arquivo Inválido!");
+        }
     }
 
     /**
      * Função para excluir um anexo pelo seu ID, recebido pela variável "id"
+     *
+     * @param id ID do anexo a ser excluído
+     * @return ResponseEntity com string confirmando a deleção
      */
     @Transactional
     @DeleteMapping("/{id}")
     public ResponseEntity<Object> deleteById(@PathVariable(value = "id") Long id) {
-        Anexo anexo = anexoService.findById(id);
-        anexoService.deleteById(id);
-        return ResponseEntity.status(HttpStatus.OK).body(anexo);
-    }
-
-    /**
-     * Função para excluir um anexo pelo seu nome, recebido pela variável "nome"
-     */
-    @Transactional
-    @DeleteMapping("/nome/{nome}")
-    public ResponseEntity<Object> deleteByNome(@PathVariable(value = "nome") String nome) {
-        try {
-            Anexo anexo = anexoService.findByNome(nome);
-            anexoService.deleteByNome(nome);
-            return ResponseEntity.status(HttpStatus.OK).body(anexo);
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.OK).body("Sem anexo!");
+        if (!anexoService.existsById(id)) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Anexo não encontrado!");
         }
+
+        anexoService.deleteById(id);
+        return ResponseEntity.status(HttpStatus.OK).body("Anexo deletado com sucesso!");
     }
 }
